requirements.txt
python-dotenv==0.20.0
selenium==4.4.3
webdriver_manager==4.0.1
